+++ server/http/restapi/api.go
@@ -1,4 +1,4 @@
-package api
+package restapi
 
 import (
 	"encoding/json"
